improved useUserVaults
