--- packages/web/src/components/FormControls/FormSelectInput/types.ts
@@ -1,5 +1,0 @@
-export interface FormSelectInputOption {
-  label: string;
-  value: string;
-  icon?: string;
-}
